Extract IContentBlockView and related classes into separate files and update namespaces
