+++ src/TextRight.Editor.Wpf/View/Text/TextLineRender.cs
@@ -5,7 +5,6 @@ using System.Windows;
 using System.Windows.Media;
 using System.Windows.Media.TextFormatting;
 using TextRight.Core.ObjectModel.Blocks.Text;
-using TextRight.Core.ObjectModel.Blocks.Text.View;
 using TextRight.Core.Utilities;
 using TextRight.Editor.Text;
